Zabbix monitoring (#2145)
* install zabbix client

* add zabbix module

* subdata

* write to config files

* restart zabbix service

* write to config only on changes, not on startup

* validate values in setdata

* Monitoring config independent from zabbix

* change zabbix version

* fix install

* use contextmanager for file handling, touch file with sudo

* update valid topics

* improve package installation

* check payload

* enable/ disable monitoring

* remove whitespace

---------

+++ runs/install_packages.sh
@@ -1,11 +1,5 @@
 #!/bin/bash
 echo "install required packages with 'apt-get'..."
-if [ ! -e zabbix-release_6.2-3+debian11_all.deb ]
-then
-	wget https://repo.zabbix.com/zabbix/6.2/raspbian/pool/main/z/zabbix-release/zabbix-release_6.2-3%2Bdebian11_all.deb
-	sudo dpkg -i zabbix-release_6.2-3+debian11_all.deb
-	echo "download"
-fi
  sudo apt-get -q update
  sudo apt-get -q -y install \
  	vim bc jq socat sshpass sudo ssl-cert mmc-utils \
@@ -16,5 +10,14 @@ fi
  	python3-pip \
  	xserver-xorg x11-xserver-utils openbox-lxde-session lightdm lightdm-autologin-greeter accountsservice \
 	chromium chromium-l10n
-sudo apt install zabbix-agent2 zabbix-agent2-plugin-*
+if [ -z "$(dpkg -l | grep zabbix-agent2)" ]
+then
+    echo "install zabbix"
+	wget -P /tmp/ https://repo.zabbix.com/zabbix/6.2/raspbian/pool/main/z/zabbix-release/zabbix-release_6.2-3%2Bdebian11_all.deb
+	sudo dpkg -i /tmp/zabbix-release_6.2-3+debian11_all.deb
+	echo "download"
+	sudo apt-get -q update
+	sudo apt install zabbix-agent2 zabbix-agent2-plugin-*
+	sudo rm /tmp/zabbix-release_6.2-3+debian11_all.deb
+fi
 echo "done"
